Fix passwords being incorrectly parsed as a conn string
Discovered this while trying to use AWS IAM auth with RDS. The auth token for
IAM authentication looks like so:

`rdspostgres.123456789012.us-west-2.rds.amazonaws.com:5432/?Action=connect&DBUser=jane_doe&X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Expires=900...`

This is meant to be used in place of the password. The issue was that pgx was
seeing that as part of the connection string and was trying to parse it, which
failed of course. Handling the password separately fixed that issue.